requirements.txt
faiss_cpu
matplotlib
numpy
Pillow
streamlit
tensorflow
torch
torchvision
torchaudio
opencv-python-headless